requirements.txt
flask
flask-cors
pymongo==4.5.0
python-dotenv